refactor: split balanceSheet and cashflowStatement out of features/reports
Each cluster was already self-contained: the table and download button in both had
exactly one consumer, their own top-level component. features/reports keeps only the
shared infrastructure (ReportsTable, its error/loader/skeleton states, and
ReportsMobileSelectionTrigger).

Both new domain names match their API routes -- balance-sheet, cashflow-statement --
since neither exists under schemas/ or hooks/features/ yet. The domain is
cashflowStatement while the component stays StatementOfCashFlow, which is a public
export and cannot be renamed.

Both domains import the shared ReportsTable* pieces from features/reports.
features/profitAndLoss already did the same, which is why those five stayed put.
