+++ src/components/features/balanceSheet/BalanceSheet/BalanceSheet.tsx
@@ -14,10 +14,10 @@ import { HeaderCol } from '@blocks/Layout/Header/HeaderCol'
 import { HeaderRow } from '@blocks/Layout/Header/HeaderRow'
 import { View } from '@blocks/Layout/View/View'
 import { ConditionalBlock } from '@components/utility/ConditionalBlock'
-import { BALANCE_SHEET_ROWS_CONFIG } from '@features/reports/BalanceSheet/constants'
-import { BalanceSheetDownloadButton } from '@features/reports/BalanceSheetDownloadButton/BalanceSheetDownloadButton'
-import { BalanceSheetTable } from '@features/reports/BalanceSheetTable/BalanceSheetTable'
-import { type BalanceSheetTableStringOverrides } from '@features/reports/BalanceSheetTable/BalanceSheetTable'
+import { BALANCE_SHEET_ROWS_CONFIG } from '@features/balanceSheet/BalanceSheet/constants'
+import { BalanceSheetDownloadButton } from '@features/balanceSheet/BalanceSheetDownloadButton/BalanceSheetDownloadButton'
+import { BalanceSheetTable } from '@features/balanceSheet/BalanceSheetTable/BalanceSheetTable'
+import { type BalanceSheetTableStringOverrides } from '@features/balanceSheet/BalanceSheetTable/BalanceSheetTable'
 import { ReportsMobileSelectionTrigger } from '@features/reports/ReportsMobileSelectionTrigger/ReportsMobileSelectionTrigger'
 import { ReportsTableErrorState } from '@features/reports/ReportsTableErrorState/ReportsTableErrorState'
 import { ReportsTableLoader } from '@features/reports/ReportsTableLoader/ReportsTableLoader'
