+++ src/views/Reports/Reports.tsx
@@ -9,15 +9,15 @@ import { type LinkingMetadata } from '@contexts/InAppLinkContext'
 import { type ReportOption, ReportsHeaderContextProvider, type ReportType } from '@contexts/ReportsHeaderContext/ReportsHeaderContext'
 import { Container } from '@blocks/Layout/Container/Container'
 import { View } from '@blocks/Layout/View/View'
+import { BalanceSheet } from '@features/balanceSheet/BalanceSheet/BalanceSheet'
+import { type BalanceSheetStringOverrides } from '@features/balanceSheet/BalanceSheet/BalanceSheet'
+import { StatementOfCashFlow } from '@features/cashflowStatement/StatementOfCashFlow/StatementOfCashFlow'
+import { type StatementOfCashFlowStringOverrides } from '@features/cashflowStatement/StatementOfCashFlow/StatementOfCashFlow'
 import { ProfitAndLoss } from '@features/profitAndLoss/ProfitAndLoss/ProfitAndLoss'
 import { type ProfitAndLossDetailedChartsStringOverrides } from '@features/profitAndLoss/ProfitAndLossDetailedCharts/ProfitAndLossDetailedCharts'
 import { type ProfitAndLossDownloadButtonStringOverrides } from '@features/profitAndLoss/ProfitAndLossDownloadButton/types'
 import { ProfitAndLossReport } from '@features/profitAndLoss/ProfitAndLossReport/ProfitAndLossReport'
 import { type ProfitAndLossTableStringOverrides } from '@features/profitAndLoss/ProfitAndLossTable/ProfitAndLossTableComponent'
-import { BalanceSheet } from '@features/reports/BalanceSheet/BalanceSheet'
-import { type BalanceSheetStringOverrides } from '@features/reports/BalanceSheet/BalanceSheet'
-import { StatementOfCashFlow } from '@features/reports/StatementOfCashFlow/StatementOfCashFlow'
-import { type StatementOfCashFlowStringOverrides } from '@features/reports/StatementOfCashFlow/StatementOfCashFlow'
 import { ReportsToggle } from '@views/Reports/ReportsToggle'
 
 import './reports.scss'
